+++ src/main/java/com/unstoppabledomains/exceptions/ns/NamingServiceException.java
@@ -1,4 +1,4 @@
-package com.unstoppabledomains.exceptions;
+package com.unstoppabledomains.exceptions.ns;
 
 public class NamingServiceException extends Exception {
   private static final long serialVersionUID = 1L;
