+++ src/main/java/com/unstoppabledomains/resolution/dns/DnsRecord.java
@@ -1,4 +1,4 @@
-package com.unstoppabledomains.util;
+package com.unstoppabledomains.resolution.dns;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
